backend/src/requirements.txt
boto==2.48.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.4
django-storages==1.6.6
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
psycopg2-binary==2.7.4
PyJWT==1.6.4
pytz==2018.4
static3==0.7.0
